황승민 : GetDigits, GetLenth함수 추가

+++ WinAPI_HSM_PrincessMaker2/GameEngineCore/GameEngineLevel.cpp
@@ -182,7 +182,7 @@ void GameEngineLevel::ActorsRender(float _DeltaTime)
 		for (int i = 0; i < DebugTexts.size(); i++)
 		{
 			HDC ImageDc = GameEngineWindow::GetDoubleBufferImage()->GetImageDC();
-			TextOutA(ImageDc, TextOutStart.ix(), TextOutStart.iy(), DebugTexts[i].c_str(), DebugTexts[i].size());
+			TextOutA(ImageDc, TextOutStart.ix(), TextOutStart.iy(), DebugTexts[i].c_str(), static_cast<int>(DebugTexts[i].size()));
 			TextOutStart.y += 20.0f;
 		}
